refactor(view): `LottoOutputView`의 책임을 여러 클래스로 분배
