+++ src/main/java/lotto/view/ErrorOutputView.java
@@ -1,0 +1,5 @@
+package lotto.view;
+
+public interface ErrorOutputView extends OutputView {
+    void printErrorMessage(Exception error);
+}
