+++ src/main/java/lotto/view/GuideOutputView.java
@@ -1,0 +1,9 @@
+package lotto.view;
+
+public interface GuideOutputView extends OutputView {
+    void printPurchaseAmountGuide();
+
+    void printWinningNumbersGuide();
+
+    void printBonusNumberGuide();
+}
