+++ src/main/java/lotto/view/LottoOutputView.java
@@ -4,18 +4,8 @@ import java.util.List;
 import lotto.dto.LottoDto;
 import lotto.dto.LottoStatisticDto;
 
-public interface LottoOutputView {
-    void printPurchaseAmountGuide();
-
-    void printWinningNumbersGuide();
-
-    void printBonusNumberGuide();
-
+public interface LottoOutputView extends OutputView {
     void printPurchasedLottos(List<LottoDto> lottos);
 
     void printResultStatistic(LottoStatisticDto statistic);
-
-    void printErrorMessage(Exception error);
-
-    void printEmptyLine();
 }
