+++ src/main/java/lotto/view/OutputView.java
@@ -1,0 +1,7 @@
+package lotto.view;
+
+public interface OutputView {
+    default void printEmptyLine() {
+        System.out.println();
+    }
+}
